Bug:  Enemy gitting stuck and spinning instead of patrolling
The "PlatformFinder" script was registring the Sword collider as an obsticale when checking above a tile. When the enemy got stuck inside of the collider, either by being placed inside of it or above it and falling in, the enemy script did not know how to handle it so ut just started spinning around.
- Quick fix: I made a new layer called "Items" and added the Sword's prefab to it.