Added visuals to the item tutorial
Added 2 steps in the part where we change the name of the item in the scene to also add a sprite to it so that the player actaully can see the item in the game. Removed 'armor' sprite from item prefab so the prefab sprite field is empty. Changed some masking to make it easier to drag the item into the scene.