Bug: Enemy falling through the floor
The bug seems to occur when the enemy collider gets pushed into the tilemap by the players collider, especially when the enemy is falling down a ledge when chasing the player. Since the player collider has its "collision detection" set to continuous, it has a higher priority to the enemy collider. This occur at an increased frequency they higher the enemy's patrol speed is, though i managed to recreate it at the a patrol speed of 4.
- Added a [Range] attribute to the "patrolSpeed" variable in the "PatrolEnemy" script. Limited to 0-10.
- Set the Enemy prefab's "collision detection" to continuous.